[misc] Move `threaded`, `Expose` form Addon to misc
