+++ module/plugins/hooks/AntiStandby.py
@@ -12,8 +12,8 @@ try:
 except ImportError:
     pass
 
-from module.plugins.internal.Addon import Addon, Expose
-from module.plugins.internal.misc import encode, fsjoin
+from module.plugins.internal.Addon import Addon
+from module.plugins.internal.misc import encode, Expose, fsjoin
 
 
 class Kernel32(object):
@@ -27,7 +27,7 @@ class Kernel32(object):
 class AntiStandby(Addon):
     __name__    = "AntiStandby"
     __type__    = "hook"
-    __version__ = "0.15"
+    __version__ = "0.16"
     __status__  = "testing"
 
     __config__ = [("activated", "bool", "Activated"                       , True ),
